test: 보너스 번호 입력 관련 예외처리 테스트 코드 구현
